requirements.txt

flask>=2.3.0
flask-cors>=4.0.0
pyjwt>=2.8.0
requests>=2.31.0